requirements.txt
###############################################################################
# pip dependent packages
###############################################################################

###############################################################################
# default ARGOS-LABS SDK module
###############################################################################
alabs.common

###############################################################################
# For your dependent modules
###############################################################################
# ...

#telegram_module
python-telegram-bot

#desktop_screenshot
pillow

